+++ platform/UI/dialogs/quit_dialog.cpp
@@ -2,6 +2,7 @@
 #include "Core/editor_project_globals.h"
 #include "Core/viewer_save.h"
 #include "UI/dialogs/quit_dialog.h"
+#include "UI/tools/palette_color_tools.h"
 #include "imgui.h"
 
 #include <stdlib.h>
@@ -18,6 +19,7 @@ void draw_quit_dialog(void)
             ImGui::Text("Save before quitting?");
             ImGui::Separator();
             if (ImGui::Button("Save && Quit", ImVec2(110, 0))) {
+                palette_color_shift_cancel_preview();
                 if (g_have_bdb && g_bdb_path[0]) bdb_save(g_bdb_path);
                 if (g_bdd_path[0]) bdd_save();
                 g_dirty = 0;